A3

+++ kern/arch/mips/vm/dumbvm.c
@@ -261,7 +261,7 @@ free_kpages(vaddr_t addr)
 
 	unsigned long page_belong = ((phaddr - (coremap_start + (coremap_npages * PAGE_SIZE)))/PAGE_SIZE);
 	unsigned long cur_page = page_belong;
-	kprintf("%s%d\n", "cur_page: ",cur_page);
+	kprintf("%s%d\n", "cur_page: ",(int)cur_page);
 	unsigned long alloced = ((int *) PADDR_TO_KVADDR(coremap_start))[cur_page];
 	unsigned long free_start = 0;
